added blog applciation changes done

+++ src/main/java/com/personalBloggingapi/pb/repository/BlogRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlogRepo extends JpaRepository<Article,Integer> {
+public interface BlogRepo extends JpaRepository<Article,Long> {
 }
